fix: collapse accounts and adjustments into ledger admin tabs (closes #207)
